chatbot-server/gemini/requirements.txt
chromadb==0.3.29
langchain==0.0.222
langchain-google-genai==1.0.3
google.generativeai==0.5.2
python-dotenv==1.0.1
pip==23.1.2
flask==2.2.3
flask_cors==4.0.0
# pydantic==1.10.7
websockets==11.0
uvicorn==0.22.0
Flask-Cors==1.10.3
unstructured==0.13.6
markdown==3.6
langchain-community==0.0.36
